+++ python/lsst/ts/IntegrationTests/configs/enabled_offline_state_transition_configs.py
@@ -45,6 +45,7 @@ yaml_string = yaml.safe_load(
     data:
     - [Scheduler:1, OFFLINE]
     - [Scheduler:2, OFFLINE]
+    - [Scheduler:3, OFFLINE]
     - [OCPS:1, OFFLINE]
     """
 )
@@ -68,10 +69,13 @@ yaml_string = yaml.safe_load(
     - [ESS:104, OFFLINE]
     - [ESS:105, OFFLINE]
     - [ESS:106, OFFLINE]
+    - [ESS:107, OFFLINE]
+    - [ESS:108, OFFLINE]
     - [ESS:201, OFFLINE]
     - [ESS:202, OFFLINE]
     - [ESS:203, OFFLINE]
     - [ESS:204, OFFLINE]
+    - [ESS:205, OFFLINE]
     - [ESS:301, OFFLINE]
     - [WeatherForecast, OFFLINE]
     """
@@ -85,14 +89,26 @@ registry["eas_enabled_offline"] = yaml.safe_dump(
 yaml_string = yaml.safe_load(
     """
     data:
-    - [Authorize, OFFLINE]
     - [Test:42, OFFLINE]
     - [ScriptQueue:1, OFFLINE]
     - [ScriptQueue:2, OFFLINE]
+    - [ScriptQueue:3, OFFLINE]
     """
 )
 
 registry["sq_enabled_offline"] = yaml.safe_dump(
+    yaml_string, explicit_start=True, canonical=True
+)
+
+# MTCS enabled_offline
+yaml_string = yaml.safe_load(
+    """
+    ignore:
+        - mtptg
+    """
+)
+
+registry["mtcs_enabled_offline"] = yaml.safe_dump(
     yaml_string, explicit_start=True, canonical=True
 )