+++ doc/version-history.rst
@@ -10,6 +10,14 @@ Version History
 .. No new work should be required in order to complete this section.
 .. Below is an example of a version history format.
 
+v0.24.0
+-------
+* Added ESS:205 to the EAS state transition configurations.
+* Added ESS:107, ESS:108, Scheduler:3 and ScriptQueue:3 CSCs
+* Set EPM:1 and MTPtg to individual state transition.
+* Added an ignore parameter for the MTPtg to the enabled_offline script.
+* Removed Authorize CSC.
+
 v0.23.0
 -------
 * Fixed the csc_state_transition.py script to add the csc_index to the Class instantiation.